+++ gen/com/kaylerrenslow/armaDialogCreator/expression/ExpressionLexer.java
@@ -1,4 +1,4 @@
-// Generated from D:/Archive/Intellij Files/Arma Tools/Arma Dialog Creator/src/com/kaylerrenslow/armaDialogCreator/expression\Expression.g4 by ANTLR 4.5.3
+// Generated from D:/Archive/Intellij Files/Arma Tools/Arma Dialog Creator/src/com/kaylerrenslow/armaDialogCreator/expression\Expression.g4 by ANTLR 4.7
 package com.kaylerrenslow.armaDialogCreator.expression;
 import org.antlr.v4.runtime.Lexer;
 import org.antlr.v4.runtime.CharStream;
@@ -11,7 +11,7 @@ import org.antlr.v4.runtime.misc.*;
 
 @SuppressWarnings({"all", "warnings", "unchecked", "unused", "cast"})
 public class ExpressionLexer extends Lexer {
-	static { RuntimeMetaData.checkVersion("4.5.3", RuntimeMetaData.VERSION); }
+	static { RuntimeMetaData.checkVersion("4.7", RuntimeMetaData.VERSION); }
 
 	protected static final DFA[] _decisionToDFA;
 	protected static final PredictionContextCache _sharedContextCache =
@@ -20,6 +20,10 @@ public class ExpressionLexer extends Lexer {
 		Plus=1, Minus=2, FSlash=3, Star=4, LParen=5, RParen=6, Min=7, Max=8, Identifier=9, 
 		IntegerLiteral=10, FloatLiteral=11, Digits=12, DecSignificand=13, DecExponent=14, 
 		HexLiteral=15, HexDigits=16, Letter=17, LetterOrDigit=18, WhiteSpace=19;
+	public static String[] channelNames = {
+		"DEFAULT_TOKEN_CHANNEL", "HIDDEN"
+	};
+
 	public static String[] modeNames = {
 		"DEFAULT_MODE"
 	};
@@ -87,6 +91,9 @@ public class ExpressionLexer extends Lexer {
 	public String getSerializedATN() { return _serializedATN; }
 
 	@Override
+	public String[] getChannelNames() { return channelNames; }
+
+	@Override
 	public String[] getModeNames() { return modeNames; }
 
 	@Override
@@ -122,7 +129,7 @@ public class ExpressionLexer extends Lexer {
 	}
 
 	public static final String _serializedATN =
-		"\3\u0430\ud6d1\u8206\uad2d\u4417\uaef1\u8d80\uaadd\2\25\u0092\b\1\4\2"+
+		"\3\u608b\ua72a\u8133\ub9ed\u417c\u3be7\u7786\u5964\2\25\u0092\b\1\4\2"+
 		"\t\2\4\3\t\3\4\4\t\4\4\5\t\5\4\6\t\6\4\7\t\7\4\b\t\b\4\t\t\t\4\n\t\n\4"+
 		"\13\t\13\4\f\t\f\4\r\t\r\4\16\t\16\4\17\t\17\4\20\t\20\4\21\t\21\4\22"+
 		"\t\22\4\23\t\23\4\24\t\24\4\25\t\25\3\2\3\2\3\3\3\3\3\4\3\4\3\5\3\5\3"+
@@ -136,12 +143,12 @@ public class ExpressionLexer extends Lexer {
 		"\2\26\3\3\5\4\7\5\t\6\13\7\r\b\17\t\21\n\23\13\25\f\27\r\31\16\33\17\35"+
 		"\20\37\21!\22#\23%\24\'\25)\2\3\2\f\4\2GGgg\4\2--//\4\2ZZzz\5\2\62;CH"+
 		"ch\6\2&&C\\aac|\4\2\2\u0101\ud802\udc01\3\2\ud802\udc01\3\2\udc02\ue001"+
-		"\7\2&&\62;C\\aac|\5\2\13\f\17\17\"\"\u009f\2\3\3\2\2\2\2\5\3\2\2\2\2\7"+
-		"\3\2\2\2\2\t\3\2\2\2\2\13\3\2\2\2\2\r\3\2\2\2\2\17\3\2\2\2\2\21\3\2\2"+
-		"\2\2\23\3\2\2\2\2\25\3\2\2\2\2\27\3\2\2\2\2\31\3\2\2\2\2\33\3\2\2\2\2"+
-		"\35\3\2\2\2\2\37\3\2\2\2\2!\3\2\2\2\2#\3\2\2\2\2%\3\2\2\2\2\'\3\2\2\2"+
-		"\3+\3\2\2\2\5-\3\2\2\2\7/\3\2\2\2\t\61\3\2\2\2\13\63\3\2\2\2\r\65\3\2"+
-		"\2\2\17\67\3\2\2\2\21;\3\2\2\2\23?\3\2\2\2\25F\3\2\2\2\27J\3\2\2\2\31"+
+		"\7\2&&\62;C\\aac|\5\2\13\f\17\17\"\"\2\u009f\2\3\3\2\2\2\2\5\3\2\2\2\2"+
+		"\7\3\2\2\2\2\t\3\2\2\2\2\13\3\2\2\2\2\r\3\2\2\2\2\17\3\2\2\2\2\21\3\2"+
+		"\2\2\2\23\3\2\2\2\2\25\3\2\2\2\2\27\3\2\2\2\2\31\3\2\2\2\2\33\3\2\2\2"+
+		"\2\35\3\2\2\2\2\37\3\2\2\2\2!\3\2\2\2\2#\3\2\2\2\2%\3\2\2\2\2\'\3\2\2"+
+		"\2\3+\3\2\2\2\5-\3\2\2\2\7/\3\2\2\2\t\61\3\2\2\2\13\63\3\2\2\2\r\65\3"+
+		"\2\2\2\17\67\3\2\2\2\21;\3\2\2\2\23?\3\2\2\2\25F\3\2\2\2\27J\3\2\2\2\31"+
 		"M\3\2\2\2\33Z\3\2\2\2\35^\3\2\2\2\37j\3\2\2\2!u\3\2\2\2#\177\3\2\2\2%"+
 		"\u0087\3\2\2\2\'\u008c\3\2\2\2)\u0090\3\2\2\2+,\7-\2\2,\4\3\2\2\2-.\7"+
 		"/\2\2.\6\3\2\2\2/\60\7\61\2\2\60\b\3\2\2\2\61\62\7,\2\2\62\n\3\2\2\2\63"+