+++ CHANGES.md
@@ -1,9 +1,9 @@
 **Added:**
-* method in HeaderFile for getting inherited assignments
+* 
 
 **Changed:**
-* made HeaderClass have a reference to its owner file
-    * this required modified the grammar _slightly_
+* more progress on HeaderToProject
+    * converting classes into ArmaControl is making good progress
 
 **Fixed:**
 *